1) Detach the shell for running the reboot command and delay it by 5 seconds
On Centos 7, the reboot happens so quickly it tears down the ssh connection and causes an unignorable connection error.
For the time being, this is a reliable working workaround, leaving a slightly sour taste in my mouth for having to copy
a helper script on the host to be rebooted.

However, it works - reliably.

2) Amend any minute timers to be in seconds to reduce reboot wait times

Especially usefully during development iterations :)

---

modified:   defaults/main.yml
new file:   files/helper/reboot-soon.sh
modified:   tasks/main.yml